FIRST COMMIT => A lot of work will have to be done from this commit. There is nothing really worth something right now, basically just a framework in progress :P. The logging framework is completed and a basic idea has been implemented for the manager of the application in the constructors area, but nothing is really something right now.
In the next commits will come an actual logging system to files (as of now the infos are logged in queues but the queues are never emptied)
There will be as well some basic work camps that are actually doing their work.
Finally, the code that is present in this commit will me commented and todos will be added for future work.